refactor: separate discount request table into a component
